netsh: profiles: use line number and regex of "Key contents" instead of keyword itself
Previously "Key contents" is used to extract passphrase from the console
netsh command. However, Microsoft internally hard-coded i18n elements in
the netsh and could not be reverted to English unless installing
language pack for another Language,
Thus it distracts for this program to parse sementic without making an
effort to translate the keyword

So, intead of keyword, use line number where it print out to the console
to provide service not being affected by i18n.

This feature closes #4
This feature closes #7